refactor : 나이 할인금액 enum의 메소드 클래스로 분리
