+++ src/test/java/nextstep/subway/path/domain/AgeDiscountCalculatorTest.java
@@ -6,14 +6,14 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-public class AgeDiscountPolicyTest {
+public class AgeDiscountCalculatorTest {
 
     @DisplayName("나이별 할인 확인 - 성인")
     @ParameterizedTest
     @CsvSource(value = { "1050:19:1050", "2550:30:2550" }, delimiter = ':')
     void 성인_나이_할인_확인(int fare, int age, int expected) {
         // when
-        int 요금 = AgeDiscountPolicy.discount(fare, age);
+        int 요금 = AgeDiscountCalculator.discount(fare, age);
 
         // then
         assertThat(요금).isEqualTo(expected);
@@ -24,7 +24,7 @@ public class AgeDiscountPolicyTest {
     @CsvSource(value = { "1050:6:350", "2550:12:1100" }, delimiter = ':')
     void 어린이_나이_할인_확인(int fare, int age, int expected) {
         // when
-        int 요금 = AgeDiscountPolicy.discount(fare, age);
+        int 요금 = AgeDiscountCalculator.discount(fare, age);
 
         // then
         assertThat(요금).isEqualTo(expected);
@@ -35,7 +35,7 @@ public class AgeDiscountPolicyTest {
     @CsvSource(value = { "1050:13:560", "2550:18:1760" }, delimiter = ':')
     void 청소년_나이_할인_확인(int fare, int age, int expected) {
         // when
-        int 요금 = AgeDiscountPolicy.discount(fare, age);
+        int 요금 = AgeDiscountCalculator.discount(fare, age);
 
         // then
         assertThat(요금).isEqualTo(expected);
